Split JuliaStackFrame into two pieces, with code being a singleton
This avoids the need for fetching the AST and copying the CodeInfo
repeatedly for the same method.

This also stashes an updated pc for each frame.
This seems more consistent with the notion that the state of all
the locals, ssavalues, etc, is dependent on the pc, so allowing
these two to diverge seems inconsistent.