Resolve world-age issues in prepare_toplevel
This reworks prepare_toplevel so that it produces a list of frames
that can be executed in order. In conjunction with another
addition, through_methoddef_or_done!, one can run each frame sequentially
up to the point of a new method definition, after which control can
return to top-level. This allows the world age to update.

While this places more burden on the top-level evaluator (one
now requires a nest of two loops), this resolves the difficulties
with previous strategies. It fixes a test previously marked as broken.