Integrate with CodeTracking and support line numbers when setting breakpoints
If the user is running an appropriate version of Revise, reported
location will update as code moves around due to edits. Because methods
can move files as well as from line-to-line, `linenumber`
should not be used except internally. (Use `whereis` instead.)

If the user is not running Revise, CodeTracking will be harmless.