A new serialization format for optimizing & executing lowered IR
This is a draft of a new tokenization of lowered IR. The main goal is
to split off from the representation currently used in Base so that
we are free to perform more significant transformations of the IR
to enable more performance optimizations.

This tackles the first step, serializing lowered code to the new
tokenized format. Currently the only things supported are serializing
and printing; execution is in draft form but never tested.