Check that the id appears in Base.loaded_modules
The problem is that if there's a file on the LOAD_PATH with the name
of the module, Base.identify_package will return a PkgId for it.
This section of the code is trying to avoid *re*-defining a module,
so we need to check that if we create a PkgId that it corresponds to a
loaded module.