Fix some invalidations from `using Images` (#430)
There are some invalidations triggered by `ColorTypes` definition of
a `length` method for `Colorant` types. The annotation in the
comprehension is what fixes it. But there are a couple of other inference
issues, and I may have been more specific in a previous annotation
than merited by reality (not quite sure why it hasn't failed...)