Try harder to find __toplevel__ stdlibs in ExprSplitter (#447)
Revise has to be able to analyze code in modules that are not in
the namespace. It turns out that differences between how discovery
works in interactive mode (which implicitly widens the Manifest to
include all the stdlibs) and test-mode (which does not) have led to
the handling discovery of toplevel modules problematic for some time:
it turns out we've been constructing "shadow" modules of the same names
as these stdlibs. Some of the exclusions in Revise's `sigtest.jl` test
file reflect this.

This causes ExprSplitter to try harder to find modules from
`Base.__toplevel__`. When we can't resolve the name in the manifest,
we also check Base.loaded_modules and see if we can find something of
that name.

This fixes a bug that just manifested (but whose roots are old) for
Revise on nightly.