fix world age issue when interpreting toplevel frame with function definitions
While debuging https://github.com/aviatesk/JET.jl/issues/104, I found
the same could happen for JuliaInterpreter.

This issue doesn't happen when interpreting an local frame, so ideally
we want to use `Base.invokelatest` only when interpreting a toplevel
frame. But it needs some refactoring on the whole interpret.jl code
base, and so I want to leave it as a future work if you're okay.