Simplify precompiles
This switches to "do a little work" for precompilation,
rather than explicitly spelling out the specific signatures.
This should be more maintainable and potentially more
exhaustive (or could be made to be so) since it
precompiles across runtime dispatch boundaries.

There are also a couple of improvements in `step_expr!`
inference that I made while trying to track down some
invalidations. These weren't the cause, but they don't
hurt.