[incomplete] changes needed to adapt to compressed line table format in Julia (#606)
This commit is incomplete, but since I don't hold back other packages in
the ecosystem, I'm going to release a version that allows for
precompilation of this package in v1.12. I've made the issues for
specific problems found during the PR and we can fix them one by one
later on.