minimum adjustments to make JuliaInterpreter work with v1.12
For now, I've made the minimal changes needed to use JuliaInterpreter
(and Revise) with v1.12. CI is still broken, so additional changes are
needed.

Additionally, while making these changes, I found that the
implementation of `JuliaInterpreter.optimize!` is quite illegal.
This is an issue that should be resolved eventually, so I've left
comments in optimize.jl regarding the known problems.