Handle `:latestworld` expr
These were added in https://github.com/JuliaLang/julia/pull/56523.
This simply adds tracking for the world age to the interpreter, but
does not use it for anything. JuliaInterpreter's current model
of world age does not match Base anyway. We should fix that eventually,
but it's probably easier to do that once https://github.com/JuliaLang/julia/pull/56509
and binding partitions are merged.