change the `recurse` interface to `AbstractInterpreter`-like interface (#683)
Align the `recurse` argument to something like the base Compiler's
`AbstractInterpreter` and make JuliaInterpreter routines overloadable
properly.

This change is quite breaking (thus bumping the minor version of this
package), but necessary to enhance the customizability of JI.
For example, it will make it easier to add changes like
JuliaDebug/JuliaInterpreter.jl#682 in a nicer way, but also should
enable better designs in packages such as Revise and JET.

- allow specifying interpreter with `@interpret`
- better error handling for `@interpret`
- better `native_call` implementation
- make `lookup` overloadable
- define `NonRecursiveInterpreter` and make `Compiled` alias to it
`Compiled` is preserved for backward compatibility.

---------

Co-authored-by: Cédric Belmant <cedric.bel@hotmail.fr>