implement a system for interpreting overlay methods (#682)
* change the `recurse` interface to `AbstractInterpreter`-like interface

Align the `recurse` argument to something like the base Compiler's
`AbstractInterpreter` and make JuliaInterpreter routines overloadable
properly.

This change is quite breaking (thus bumping the minor version of this
package), but necessary to enhance the customizability of JI.
For example, it will make it easier to add changes like
JuliaDebug/JuliaInterpreter.jl#682 in a nicer way, but also should
enable better designs in packages such as Revise and JET.

* implement a system for interpreting overlay methods

Implemented, being inspired by JuliaDebug/JuliaInterpreter.jl#680,
thinking it might be useful in the near future for JET's use case.

The overlay interpretertation is enabled by overloading
`JuliaInterpreter.method_table(recurse)`, but maybe the entire
`recurse`-overload mechanism itself needs to be aligned with the
`AbstractInterpreter` design and properly organized.

Since this interface will also be needed for JET,
I'll probably work on it soon.