Update examples and infrastructure in documentation (#704)
* Update syntax highlight tags in example code blocks

Change the language from `julia` to `julia-repl`

* Update outputs in code block examples

* Use `@repl` for REPL session example

The code block calls `Meta.lower()` which includes styled output, and
its easier to see when Documenter.jl itself renders it rather than
having the output already in the Markdown file and trying to syntax
highlight it.

* Add julia compat bounds (`~1.10`) to docs/Project.toml

Currently it's handled by Manifest.toml, but Project.toml is easier to
check against and it lets devs know that the docs should be built with a
particular version of julia.

There should also probably be compat bounds on the other docs
dependencies.